// backend/src/modules/production-orders/dto/update-production-order-hardness-check.dto.ts
export class UpdateProductionOrderHardnessCheckDto {
  requirement?: string | null;
  dosage_form_stage?: string | null;
  unit_1_hardness?: number | string | null;
  unit_2_hardness?: number | string | null;
  unit_3_hardness?: number | string | null;
  unit_4_hardness?: number | string | null;
  unit_5_hardness?: number | string | null;
  unit_6_hardness?: number | string | null;
  unit_7_hardness?: number | string | null;
  unit_8_hardness?: number | string | null;
  unit_9_hardness?: number | string | null;
  unit_10_hardness?: number | string | null;
  unit?: string | null;
}